+++ src/llm_api_adapter/adapters/google/__init__.py
@@ -1,0 +1,5 @@
+"""Google adapter package."""
+
+from .adapter import GoogleAdapter
+
+__all__ = ["GoogleAdapter"]
